#28: Move link model
