+++ bowler-kernel/kinematics/src/main/kotlin/com/neuronrobotics/bowlerkernel/kinematics/limb/link/model/LinkScriptData.kt
@@ -14,11 +14,11 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.neuronrobotics.bowlerkernel.kinematics.limb.model
+package com.neuronrobotics.bowlerkernel.kinematics.limb.link.model
 
-import com.neuronrobotics.bowlerkernel.kinematics.limb.link.LinkType
+import com.neuronrobotics.bowlerkernel.kinematics.motion.model.ControllerSpecification
 
-data class LinkConfigurationData(
-    val type: LinkType,
-    val dhParamData: DhParamData
+data class LinkScriptData(
+    val jointAngleController: ControllerSpecification,
+    val inertialStateEstimator: ControllerSpecification
 )
